requirements.txt
pandas
requests
pybaseball
pyYAML
python-dotenv
json
weasyprint
beautifulsoup4